[Rename] : move package envdata
