+++ src/main/java/com/saeromteo/app/controller/NewsController.java
@@ -1,0 +1,5 @@
+package com.saeromteo.app.controller;
+
+public class NewsController {
+	
+}
